performPlugin: Use the correct match string
I had been using the regex instead of the match result. Do another regex
match to correctly output our SHA instead of the match regex.

I had tried putting the `=~` match in another function and using it from
here and in the list comprehension in `matchPlugins`, but then found out
that, of course, the return types didn't agree (`matchPlugins` needs a
`Bool` and `performPlugin` needs a `String`). Ended up dispensing with
the extra function and doing the match both times instead.