Install 'dbmigrations-sqlite'
Use this for database migrations. Seems to be the right tool for
database migrations in Haskell. Originally I had thought about doing it
manually with 'up' and 'down' SQL files and some kind of Makefile or
script to run them, but then realised that, of course, you have to keep
track of which migration you're on, and figured, well, might want to
outsource that to another program.

I had originally written my migration in Mark Bates' pop/soda format,
but will have to convert that for 'dbmigrations'.

Thankfully got a little help from
https://github.com/basti1302/elm-lang-de#development-setup
on how to actually install the 'dbmigrations-sqlite' package using
Stack. Put the installation command into a new "setup.sh" file to make
it easily repeatable.

Needed to add a bunch of lines to the `extra-deps` section of
"stack.yaml" in order to be able to properly install the package.