Add `Message` types
Add a few types to collect properties of a message, including message
body, channel posted to, and sender.

Need something like this for the GitHub Commit plugin, where we need
access to the message to get the SHA, and we need the channel to get the
repo URL from the database.