Set up basic IRC support
Add the 'irc-client' package to facilitate communication over IRC.

Copy the example from:
http://hackage.haskell.org/package/irc-client-0.4.4.4/docs/Network-IRC-Client.html

which connects to a network. We specify a channel to join, and what do
you know, it works! Pretty cool.

Commented out the code that runs the GitHub Commit plugin for now while
testing this.

Add the `OverloadedStrings` extension so we can write string literals
but have them converted into the appropriate `Data.Text` and
`Data.ByteString` types required by the IRC library.