stack.yaml: Turn off `-support_aesni` when compiling 'cryptonite'
The 'irc-client' package depends on 'cryptonite'. I was getting this
compilation error from 'cryptonite' on my machine:

    [111 of 111] Compiling Crypto.Cipher.AES ( Crypto/Cipher/AES.hs, .stack-work/dist/x86_64-osx/Cabal-1.24.2.0/build/Crypto/Cipher/AES.o )
    cc1: error: unrecognized command line option "-maes"
    `gcc' failed in phase `C Compiler'. (Exit code: 1)

The package recommends disabling AESNI via a compile flag to get around
this error
(https://github.com/haskell-crypto/cryptonite#known-building-issues):

    cabal configure --flag='-support_aesni'

Add 'cryptonite' to our `extra-deps` in order to be able to set a flag
on it. Turn off the `-support_aesni` flag in order to facilitate
compilation.