Lib.hs: Test chat message listening and posting
Add an event handler that gets called when a message is posted to a
channel the bot is on. When a message is posted, the bot will send a
chat message containing the text "test".

This tests out the message handling and posting mechanism, and gives us
a place to build off of to allow the bot to communicate.