Set up i18n
en.msg:
* Remove comment lines because apparently comments aren't allowed by the
  parser
* Change data type to `T.Text` to match the import from the `I18n`
  module

GitHubCommit.hs:
Import the `I18n` module because we'll be testing it in context in a
bit, but for now just want to get the module code to compile.

I18n.hs:
Follow the example on http://www.yesodweb.com/book/internationalization
to make a function for proper internationalisation.