Try to get I18n working
An unsuccessful attempt at getting i18n working through Shakespeare. Had
a little help from:

https://gist.github.com/nicolashery/f87467fb37da2b00cec1eed028f51e60

Unfortunately, I can't figure out how to get this working. Even before,
without the locale stuff and defining a custom `translate` function,
when I tried using Shakespeare's `_{}` function, I was, and still am,
getting this error:

    src/Plugin/GitHubCommit.hs:38:29: error:
        Data constructor not in scope: MsgGitHubCommitRepoURLNotFound

Yes, obviously, that makes sense. But how do I get it to be in scope if
it isn't even in my code? What?

I'm giving up on Shakespeare for i18n. It was a nice idea, I liked the
fact that strings were stored in external files, but I don't know how to
do this, I can't find the right resources online to use this in a
non-web non-template context, and it's becoming a pain. Going to get rid
of all this code and just use regular Haskell data types to do it as
described in
https://wiki.haskell.org/Internationalization_of_Haskell_programs