Add Factorial plugin
This plugin calculates factorials. It's not super optimised, but works
for small inputs.

Adds the 'text-show' plugin to convert the resulting integer to a
`Data.Text` to be returned by the plugin action.

At first I had tried /^\d+!$/ for the regex match, but that produced
this error:

    sorbot/src/Plugin/Factorial.hs:14:24: error:
        lexical error in string/character literal at character 'd'

Tried to fix that by using /^\\d+!$/, but that didn't match anything.
Then tried a long-form character class, /^[:digit:]+!$/, which TDFA
implies it supports, but a quick look at the code seems to indicate that
those parts were removed or commented out. Not sure what was going on
with that. So instead went for the remaining solution, a `0-9` character
class.