Plugin: Add `queryOnly` field
A new field that says whether this plugin should only respond via a
private query message to the user instead of responding on the channel
the message was sent from.

This is needed for the Help plugin, which shouldn't flood channels with
lots of extraneous output. Instead, the Help plugin should send the list
of commands directly to the user.

Since most of the time we don't want this behaviour, encode a default of
`False` on the field so that most plugins don't have to define it
manually. This necessitates changing the constructors to use the default
`Plugin` instead.