CliOptions: Parse the `language` option into a `Locale`
Instead of parsing `--language` as a string, parse it as a real `Locale`
type. This means the parsing is taken care of for us at the CLI option
handling stage instead of later on.

We can thus use the language value more quickly and easily, passing it
to the `translate` function, which takes a `Locale`.