GitHubCommit: Make localisation work
Get the language from the command line options and use it to determine
which localised string to return on error. Wooo!! it works! So cool.

Add French translations for our existing strings.

The code to get the language option is pretty messy for the moment. I
want to factor it out into its own separate function. But at least it's
working now.