Create `Bot` type
This type will be used as an application-level monad, enabling us to
transmit configuration coming from CLI arguments (`Options`) to any
function that needs it.

The immediate benefit is that we'll be able to easily transmit the
specified language to functions to allow them to use the right
translation string.