GitHubCommit: Try to add `Bot` monad
This is an attempt at wrapping the functions in this plugin with the new
`Bot` monad. It didn't go very well.

These changes are the result of some initial code and then a lot of
fiddling to try to resolve compiler errors, many of which are still
present.

The ultimate goal is to be able to wrap the functions in `Bot` so that
we can use the language configuration option to determine which
translation to use.

Thanks to 'dmwit' on Freenode#haskell for looking at my code and giving
me suggestions for my type errors.