Bot(privmsgFromPlugin): Try to use the `Bot` monad instead of `IO`
Convert this function and `Plugins(performPlugin)` to use a
`Bot`-wrapped plugin.

Still having a lot of trouble with this, unable to get past
monad-related compilation errors. The plan is to leave this as-is for
now and attack the problem from a different place, looking at the `main`
function next, applying our plugins with `runBot`, and then passing
those monad-wrapped plugins to `connectIRC`. Once that's taken care of,
we'll see what we need to modify in this middle strata of the stack.